fix: address CodeRabbit review findings for HC-SR04 IRQ mode (round 6)
Fix uint32 overflow in hcsr04_hal_get_time_us_isr by promoting to
uint64_t before multiply (65535*1000000 exceeds UINT32_MAX). Add
duration sanity check returning k_rx_err_range_check_failed when
computed value exceeds CMT2 wrap period. Extract duplicated
arm-ISR/trigger/dispatch logic into internal_trigger_and_measure
helper. Add k_hcsr04_irq_priority_min/max to public enum and remove
duplicate local constants. Add Max Range row to mode comparison table,
fix rx_hcsr04_t struct alignment, enhance k_cmt2_wrap_us clock
dependency docs, add @see disarm cross-reference to isr_start, cast
enum constants in @code examples, and document stateless mock behavior.

+++ e2-studio-star-rx72n-firmware/libs/rx_hcsr04/src/rx_hcsr04_hal_hw.c
@@ -1135,7 +1135,7 @@ uint32_t hcsr04_hal_get_time_us(void)
 
   if (internal_time_mutex_init() != k_rx_ok) {
     timer_hz = k_pclkb_hz / k_cmt2_divider;
-    return (uint32_t)((cmt2()->cmcnt * k_us_per_second) / timer_hz);
+    return (uint32_t)(((uint64_t)cmt2()->cmcnt * k_us_per_second) / timer_hz);
   }
 
   /* Protect static variables from concurrent access */
@@ -1143,7 +1143,7 @@ uint32_t hcsr04_hal_get_time_us(void)
   if (mutex_status != TX_SUCCESS) {
     /* Fallback: return current counter value without overflow tracking */
     timer_hz = k_pclkb_hz / k_cmt2_divider;
-    return (uint32_t)((cmt2()->cmcnt * k_us_per_second) / timer_hz);
+    return (uint32_t)(((uint64_t)cmt2()->cmcnt * k_us_per_second) / timer_hz);
   }
 
   current_counter = cmt2()->cmcnt;
@@ -1213,5 +1213,5 @@ uint32_t hcsr04_hal_get_time_us_isr(void)
   const uint32_t timer_hz      = k_pclkb_hz / k_cmt2_divider;
   const uint16_t current_count = cmt2()->cmcnt;
 
-  return (uint32_t)(((uint32_t)current_count * k_us_per_second) / timer_hz);
+  return (uint32_t)(((uint64_t)current_count * k_us_per_second) / timer_hz);
 }
